Renamed graphqlQuery to queries
